+++ cmd/calyptia/get_resource_profile_test.go
@@ -93,7 +93,7 @@ func Test_newCmdGetResourceProfiles(t *testing.T) {
 		cmd := newCmdGetResourceProfiles(configWithMock(&ClientMock{
 			ResourceProfilesFunc: func(ctx context.Context, aggregatorID string, params cloud.ResourceProfilesParams) (cloud.ResourceProfiles, error) {
 				wantNoEq(t, nil, params.Last)
-				wantEq(t, uint64(2), *params.Last)
+				wantEq(t, uint(2), *params.Last)
 				return want, nil
 			},
 		}))
